Long Overdue Update to WP5.7

+++ wp-content/plugins/wpforms/src/Pro/Admin/Entries/Export/File.php
@@ -2,8 +2,7 @@
 
 namespace WPForms\Pro\Admin\Entries\Export;
 
-use \Goodby\CSV\Export\Standard\Exporter;
-use \Goodby\CSV\Export\Standard\ExporterConfig;
+use Goodby\CSV\Export\Standard\CsvFileObject;
 
 /**
  * File-related routines.
@@ -42,9 +41,9 @@ class File {
 	 */
 	public function hooks() {
 
-		add_action( 'wpforms_tools_init', array( $this, 'entries_export_download_file' ) );
-		add_action( 'wpforms_tools_init', array( $this, 'single_entry_export_download_file' ) );
-		add_action( 'admin_init', array( $this, 'remove_old_export_files' ) );
+		add_action( 'wpforms_tools_init', [ $this, 'entries_export_download_file' ] );
+		add_action( 'wpforms_tools_init', [ $this, 'single_entry_export_download_file' ] );
+		add_action( Export::TASK_CLEANUP, [ $this, 'remove_old_export_files' ] );
 	}
 
 	/**
@@ -52,21 +51,75 @@ class File {
 	 *
 	 * @since 1.5.5
 	 *
-	 * @param array $export_data  Export data array.
 	 * @param array $request_data Request data array.
 	 */
-	public function write_csv( $export_data, $request_data ) {
+	public function write_csv( $request_data ) {
 
 		$export_file = $this->get_tmpfname( $request_data );
 
-		// Include Exporter.
-		require_once WPFORMS_PLUGIN_DIR . 'vendor/autoload.php';
+		if ( empty( $export_file ) ) {
+			return;
+		}
 
-		$config = new ExporterConfig();
-		$config->setDelimiter( $this->export->configuration['csv_export_separator'] );
-		$config->setFileMode( 'a' );
-		$exporter = new Exporter( $config );
-		$exporter->export( $export_file, $export_data );
+		$csv       = new CsvFileObject( $export_file, 'a' );
+		$enclosure = '"';
+
+		$csv->fputcsv( $request_data['columns_row'], $this->export->configuration['csv_export_separator'], $enclosure );
+
+		for ( $i = 1; $i <= $request_data['total_steps']; $i ++ ) {
+
+			$entries = wpforms()->entry->get_entries( $request_data['db_args'] );
+
+			foreach ( $this->export->ajax->get_entry_data( $entries ) as $entry ) {
+
+				$csv->fputcsv( $entry, $this->export->configuration['csv_export_separator'], $enclosure );
+			}
+
+			$request_data['db_args']['offset'] = $i * $this->export->configuration['entries_per_step'];
+		}
+
+		$csv->fflush();
+	}
+
+	/**
+	 * Write data to .xlsx file.
+	 *
+	 * @since 1.6.5
+	 *
+	 * @param array $request_data Request data array.
+	 */
+	public function write_xlsx( $request_data ) {
+
+		$export_file = $this->get_tmpfname( $request_data );
+
+		if ( empty( $export_file ) ) {
+			return;
+		}
+
+		$writer     = new \XLSXWriter();
+		$sheet_name = 'WPForms';
+
+		$widths = [];
+
+		foreach ( $request_data['columns_row'] as $header ) {
+			$widths[] = strlen( $header ) + 2;
+		}
+
+		$writer->writeSheetHeader( $sheet_name, array_fill_keys( $request_data['columns_row'], 'string' ), [ 'widths' => $widths ] );
+
+		for ( $i = 1; $i <= $request_data['total_steps']; $i ++ ) {
+
+			$entries = wpforms()->entry->get_entries( $request_data['db_args'] );
+
+			foreach ( $this->export->ajax->get_entry_data( $entries ) as $entry ) {
+
+				$writer->writeSheetRow( $sheet_name, $entry, [ 'wrap_text' => true ] );
+			}
+
+			$request_data['db_args']['offset'] = $i * $this->export->configuration['entries_per_step'];
+		}
+
+		$writer->writeToFile( $export_file );
 	}
 
 	/**
@@ -78,24 +131,25 @@ class File {
 	 */
 	public function get_tmpdir() {
 
-		$uploads              = wp_upload_dir();
-		$wpforms_uploads_root = trailingslashit( $uploads['basedir'] ) . 'wpforms';
+		$upload_dir  = wpforms_upload_dir();
+		$upload_path = $upload_dir['path'];
 
-		// Apply filter to allow redefine tmp directory.
-		$custom_uploads_root = apply_filters( 'wpforms_upload_root', $wpforms_uploads_root );
-		if ( wp_is_writable( $custom_uploads_root ) ) {
-			$wpforms_uploads_root = $custom_uploads_root;
+		$export_path = trailingslashit( $upload_path ) . 'export';
+		if ( ! file_exists( $export_path ) ) {
+			wp_mkdir_p( $export_path );
 		}
 
-		$export_dir = trailingslashit( $wpforms_uploads_root ) . 'export';
-		if ( ! file_exists( $export_dir ) ) {
-			wp_mkdir_p( $export_dir );
-		}
+		// Check if the .htaccess exists in the upload directory, if not - create it.
+		wpforms_create_upload_dir_htaccess_file();
+
+		// Check if the index.html exists in the directories, if not - create it.
+		wpforms_create_index_html_file( $upload_path );
+		wpforms_create_index_html_file( $export_path );
 
 		// Normalize slashes for Windows.
-		$export_dir = wp_normalize_path( $export_dir );
+		$export_path = wp_normalize_path( $export_path );
 
-		return $export_dir;
+		return $export_path;
 	}
 
 	/**
@@ -108,6 +162,10 @@ class File {
 	 * @return string Temporary file full pathname.
 	 */
 	public function get_tmpfname( $request_data ) {
+
+		if ( empty( $request_data ) ) {
+			return '';
+		}
 
 		$export_dir  = $this->get_tmpdir();
 		$export_file = $export_dir . '/' . sanitize_key( $request_data['request_id'] );
@@ -132,6 +190,47 @@ class File {
 		header( 'Content-Type: text/csv' );
 		header( 'Content-Disposition: attachment; filename=' . $file_name );
 		header( 'Content-Transfer-Encoding: binary' );
+	}
+
+	/**
+	 * Output the file.
+	 *
+	 * @since 1.6.0.2
+	 *
+	 * @param array $request_data Request data.
+	 *
+	 * @throws \Exception In case of file error.
+	 */
+	public function output_file( $request_data ) {
+
+		$export_file = $this->get_tmpfname( $request_data );
+
+		if ( empty( $export_file ) ) {
+			throw new \Exception( $this->export->errors['unknown_request'] );
+		}
+
+		clearstatcache( true, $export_file );
+
+		if ( ! is_readable( $export_file ) || is_dir( $export_file ) ) {
+			throw new \Exception( $this->export->errors['file_not_readable'] );
+		}
+
+		if ( @filesize( $export_file ) === 0 ) { //phpcs:ignore
+			throw new \Exception( $this->export->errors['file_empty'] );
+		}
+
+		$entry_suffix = ! empty( $request_data['db_args']['entry_id'] ) ? '-entry-' . $request_data['db_args']['entry_id'] : '';
+
+		$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . $entry_suffix . '-' . current_time( 'Y-m-d-H-i-s' ) . '.' . $request_data['type'];
+
+		$this->http_headers( $file_name );
+
+		readfile( $export_file ); // phpcs:ignore
+
+		// Schedule clean up.
+		$this->schedule_remove_old_export_files();
+
+		exit;
 	}
 
 	/**
@@ -167,23 +266,7 @@ class File {
 			// Get stored request data.
 			$request_data = get_transient( 'wpforms-tools-entries-export-request-' . $args['request_id'] );
 
-			$export_file = $this->get_tmpfname( $request_data );
-
-			clearstatcache( true, $export_file );
-
-			if ( ! is_readable( $export_file ) ) {
-				throw new \Exception( $this->export->errors['file_not_readable'] );
-			}
-
-			if ( @filesize( $export_file ) === 0 ) { //phpcs:ignore
-				throw new \Exception( $this->export->errors['file_empty'] );
-			}
-
-			$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . '-' . date( 'Y-m-d-H-i-s' ) . '.csv';
-			$this->http_headers( $file_name );
-
-			readfile( $export_file ); // phpcs:ignore
-			exit;
+			$this->output_file( $request_data );
 
 		} catch ( \Exception $e ) {
 			// phpcs:disable
@@ -252,29 +335,13 @@ class File {
 
 			$this->export->ajax->request_data = $request_data;
 
-			// Get export data.
-			$export_data = $this->export->ajax->get_data();
-
-			// Writing to csv file.
-			$this->write_csv( $export_data, $request_data );
-
-			$export_file = $this->get_tmpfname( $request_data );
-
-			clearstatcache( true, $export_file );
-
-			if ( ! is_readable( $export_file ) ) {
-				throw new \Exception( $this->export->errors['file_not_readable'] );
+			if ( empty( $request_data['type'] ) || $request_data['type'] === 'csv' ) {
+				$this->write_csv( $request_data );
+			} elseif ( $request_data['type'] === 'xlsx' ) {
+				$this->write_xlsx( $request_data );
 			}
 
-			if ( @filesize( $export_file ) === 0 ) { //phpcs:ignore
-				throw new \Exception( $this->export->errors['file_empty'] );
-			}
-
-			$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . '-entry-' . $request_data['db_args']['entry_id'] . '-' . date( 'Y-m-d-H-i-s' ) . '.csv';
-			$this->http_headers( $file_name );
-
-			readfile( $export_file ); // phpcs:ignore
-			exit;
+			$this->output_file( $request_data );
 
 		} catch ( \Exception $e ) {
 
@@ -288,19 +355,34 @@ class File {
 	}
 
 	/**
-	 * Garbage clearing.
-	 * Actually we need to clear only temporary files
-	 * because transients clears automagically.
+	 * Register a cleanup task to remove temp export files.
 	 *
-	 * TODO: rewrite to use wp-cron or actions scheduler.
+	 * @since 1.6.5
+	 */
+	public function schedule_remove_old_export_files() {
+
+		$tasks = wpforms()->get( 'tasks' );
+
+		if ( ! empty( $tasks->is_scheduled( Export::TASK_CLEANUP ) ) ) {
+			return;
+		}
+
+		$tasks->create( Export::TASK_CLEANUP )
+		      ->recurring( time() + 60, $this->export->configuration['request_data_ttl'] )
+		      ->params()
+		      ->register();
+	}
+
+	/**
+	 * Garbage clearing.
+	 * We need to clear only temporary files because transients are cleared automagically.
 	 *
 	 * @since 1.5.5
+	 * @since 1.6.5 Stop using transients. Now we use ActionScheduler for this task.
 	 */
 	public function remove_old_export_files() {
 
-		if ( (bool) get_transient( 'wpforms_entries_export_tmpdata_cleared' ) ) {
-			return;
-		}
+		clearstatcache();
 
 		$files = glob( $this->get_tmpdir() . '/*' );
 		$now   = time();
@@ -308,12 +390,11 @@ class File {
 		foreach ( $files as $file ) {
 			if (
 				is_file( $file ) &&
+				pathinfo( $file, PATHINFO_BASENAME ) !== 'index.html' &&
 				( $now - filemtime( $file ) ) > $this->export->configuration['request_data_ttl']
 			) {
 				unlink( $file );
 			}
 		}
-
-		set_transient( 'wpforms_entries_export_tmpdata_cleared', 'yes', HOUR_IN_SECONDS );
 	}
 }
